The spring container manages the lifecycle of the beans and injects dependencies.
