+++ src/main/resources/myApplicationContext.xml
@@ -8,9 +8,15 @@
     http://www.springframework.org/schema/context/spring-context.xsd">
 
 
-    <bean id="idAnimal" class="Pet.Dog"></bean>
+    <bean id="idAnimal" class="Pet.Cat"></bean>
     <bean id="people" class="People.Superman"></bean>
 
 
+    <bean id = "empl"
+          class="DependencyInjection.Employee">
+        <constructor-arg ref="idAnimal"/>
+
+
+    </bean>
 
 </beans>
